requirements.txt
langchain==0.1.16
pypdf==4.2.0
python-docx==1.1.2
replicate==0.25.2
streamlit==1.33.0
boto3==1.34.23
